requirements.txt
aiohttp==3.9.4
aiosignal==1.3.1
asyncio==3.4.3
attrs==23.2.0
discord==2.3.2
discord.py==2.3.2
frozenlist==1.4.1
idna==3.7
logging==0.4.9.6
multidict==6.0.5
python-dotenv==1.0.1
setuptools==69.5.1
wheel==0.43.0
yarl==1.9.4
